requirements.txt
asgiref==3.7.1
certifi==2023.5.7
cryptography==41.0.1
Django==4.2.1
django-allauth==0.54.0
django-ckeditor==6.6.1
django-js-asset==2.1.0
django-stronghold==0.4.0
gunicorn==21.2.0
oauthlib==3.2.2
packaging==23.1
PyJWT==2.7.0
python-slugify==8.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
urllib3==2.0.2
whitenoise==6.5.0
